pyproject.toml
[tool.poetry]
name = "parser-bb"
version = "0.1.0"
description = ""
authors = ["halone <haloneprog@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9  "
botasaurus = "^3.2.26"
loguru = "^0.7.2"
beautifulsoup4 = "^4.12.3"
pyjwt = "^2.8.0"
pandas = "^2.2.1"
xlwt = "^1.3.0"
openpyxl = "^3.1.2"
icecream = "^2.1.3"
jupyter = "^1.0.0"
requests = "^2.31.0"
tqdm = "^4.66.2"
sqlalchemy = "^2.0.27"
psycopg2-binary = "^2.9.9"
aiohttp = "^3.9.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "main:start"
